Refactor admin page using modular components
