+++ app/admin/page.js
@@ -26,6 +26,25 @@ import {
 } from "../../lib/firebaseService"
 import { collection, getDocs, deleteDoc } from "firebase/firestore"
 import { db } from "../../lib/firebase"
+import ProductManagement from "../../components/admin/ProductManagement"
+import CDManagement from "../../components/admin/CDManagement"
+import ImageManagement from "../../components/admin/ImageManagement"
+import ProductPageManagement from "../../components/admin/ProductPageManagement"
+import {
+  authContainerStyle,
+  authFormStyle,
+  containerStyle,
+  headerStyle,
+  headerActionsStyle,
+  titleStyle,
+  tabsStyle,
+  tabStyle,
+  activeTabStyle,
+  errorBannerStyle,
+  loadingBannerStyle,
+  backButtonStyle,
+  logoutButtonStyle
+} from "../../components/admin/adminStyles"
 
 export default function AdminPage() {
   const [isAuthenticated, setIsAuthenticated] = useState(false)
@@ -40,9 +59,6 @@ export default function AdminPage() {
   const [publicImages, setPublicImages] = useState([])
   const [productPages, setProductPages] = useState([])
   const [activeTab, setActiveTab] = useState("products")
-  const [editingProduct, setEditingProduct] = useState(null)
-  const [editingCD, setEditingCD] = useState(null)
-  const [editingProductPage, setEditingProductPage] = useState(null)
   
   const router = useRouter()
 
@@ -566,985 +582,42 @@ export default function AdminPage() {
             </button>
           </div>
 
-      {/* 商品管理 */}
+
+
       {activeTab === "products" && (
-        <div style={contentStyle}>
-          <div style={sectionHeaderStyle}>
-            <h2>商品管理</h2>
-            <button onClick={handleCreateProduct} style={createButtonStyle}>
-              新增商品
-            </button>
-          </div>
-          
-          <div style={gridStyle}>
-            {products.map(product => (
-              <div key={product.id} style={cardStyle}>
-                <img 
-                  src={product.image} 
-                  alt={product.title}
-                  style={cardImageStyle}
-                  onError={(e) => {
-                    e.target.src = '/placeholder.svg'
-                  }}
-                />
-                <div style={cardContentStyle}>
-                  <h3>{product.title}</h3>
-                  <p>价格: ¥{product.price}</p>
-                  <p>状态: {product.visible ? '显示' : '隐藏'}</p>
-                  <div style={cardActionsStyle}>
-                    <button 
-                      onClick={() => setEditingProduct(product)}
-                      style={editButtonStyle}
-                    >
-                      编辑
-                    </button>
-                    <button 
-                      onClick={() => handleDeleteProduct(product.id)}
-                      style={deleteButtonStyle}
-                    >
-                      删除
-                    </button>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
+        <ProductManagement
+          products={products}
+          publicImages={publicImages}
+          onCreate={handleCreateProduct}
+          onUpdate={handleUpdateProduct}
+          onDelete={handleDeleteProduct}
+        />
       )}
 
-      {/* CD管理 */}
       {activeTab === "cds" && (
-        <div style={contentStyle}>
-          <div style={sectionHeaderStyle}>
-            <h2>CD轮播管理</h2>
-            <button onClick={handleCreateCD} style={createButtonStyle}>
-              新增CD
-            </button>
-          </div>
-          
-          <div style={gridStyle}>
-            {cds.map(cd => (
-              <div key={cd.id} style={cardStyle}>
-                <img 
-                  src={cd.image} 
-                  alt={cd.title}
-                  style={cardImageStyle}
-                  onError={(e) => {
-                    e.target.src = '/placeholder.svg'
-                  }}
-                />
-                <div style={cardContentStyle}>
-                  <h3>{cd.title}</h3>
-                  <p>链接: {cd.productLink}</p>
-                  <p>状态: {cd.visible ? '显示' : '隐藏'}</p>
-                  <div style={cardActionsStyle}>
-                    <button 
-                      onClick={() => setEditingCD(cd)}
-                      style={editButtonStyle}
-                    >
-                      编辑
-                    </button>
-                    <button 
-                      onClick={() => handleDeleteCD(cd.id)}
-                      style={deleteButtonStyle}
-                    >
-                      删除
-                    </button>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
+        <CDManagement
+          cds={cds}
+          publicImages={publicImages}
+          onCreate={handleCreateCD}
+          onUpdate={handleUpdateCD}
+          onDelete={handleDeleteCD}
+        />
       )}
 
-      {/* 图片管理 */}
       {activeTab === "images" && (
-        <div style={contentStyle}>
-          <div style={sectionHeaderStyle}>
-            <h2>图片管理</h2>
-            <label style={uploadButtonStyle}>
-              上传图片
-                  <input
-                type="file" 
-                accept="image/*" 
-                onChange={handleImageUpload}
-                style={{ display: 'none' }}
-                  />
-            </label>
-                </div>
-
-          <h3>Firebase存储的图片</h3>
-          <div style={imageGridStyle}>
-            {images.map(image => (
-              <div key={image.id} style={imageCardStyle}>
-                <img src={image.url} alt={image.name} style={imageStyle} />
-                <p>{image.name}</p>
-                <button 
-                  onClick={() => navigator.clipboard.writeText(image.url)}
-                  style={copyButtonStyle}
-                >
-                  复制链接
-                </button>
-              </div>
-            ))}
-                </div>
-
-          <h3>Public文件夹中的图片</h3>
-          <div style={imageGridStyle}>
-            {publicImages.map(imagePath => (
-              <div key={imagePath} style={imageCardStyle}>
-                <img src={imagePath} alt={imagePath} style={imageStyle} />
-                <p>{imagePath}</p>
-                <button 
-                  onClick={() => navigator.clipboard.writeText(imagePath)}
-                  style={copyButtonStyle}
-                >
-                  复制路径
-                </button>
-              </div>
-            ))}
-          </div>
-        </div>
+        <ImageManagement
+          images={images}
+          publicImages={publicImages}
+          onUpload={handleImageUpload}
+        />
       )}
 
-      {/* 产品页面管理 */}
       {activeTab === "product-pages" && (
-        <div style={contentStyle}>
-          <div style={sectionHeaderStyle}>
-            <h2>产品页面内容管理</h2>
-            <button onClick={handleCreateProductPage} style={createButtonStyle}>
-              新增页面
-            </button>
-            <p style={{ fontSize: '14px', color: '#666', marginTop: '8px' }}>
-              管理各个产品页面的详细内容，包括标题、描述、曲目列表等
-            </p>
-            <p style={{ fontSize: '12px', color: '#999', marginTop: '4px' }}>
-              可用图片: {publicImages.length} 张
-            </p>
-          </div>
-          
-          <div style={gridStyle}>
-            {productPages.map(page => (
-              <div key={page.id} style={cardStyle}>
-                <img 
-                  src={page.image} 
-                  alt={page.title}
-                  style={cardImageStyle}
-                  onError={(e) => {
-                    e.target.src = '/placeholder.svg'
-                  }}
-                />
-                <div style={cardContentStyle}>
-                  <h3>{page.title}</h3>
-                  <p>价格: {page.price}</p>
-                  <p>曲目数: {page.trackList?.length || 0}</p>
-                  <p>页面ID: {page.id}</p>
-                  <div style={cardActionsStyle}>
-                    <button 
-                      onClick={() => setEditingProductPage(page)}
-                      style={editButtonStyle}
-                    >
-                      编辑页面内容
-                    </button>
-                    <a 
-                      href={`/products/${page.id}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      style={{ 
-                        ...editButtonStyle, 
-                        textDecoration: 'none',
-                        display: 'inline-block',
-                        marginLeft: '8px'
-                      }}
-                    >
-                      预览页面
-                    </a>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* 编辑商品模态框 */}
-      {editingProduct && (
-        <ProductEditModal 
-          product={editingProduct}
-          onSave={handleUpdateProduct}
-          onCancel={() => setEditingProduct(null)}
-          availableImages={publicImages}
+        <ProductPageManagement
+          pages={productPages}
+          publicImages={publicImages}
+          onCreate={handleCreateProductPage}
+          onUpdate={handleUpdateProductPage}
         />
       )}
 
-      {/* 编辑CD模态框 */}
-      {editingCD && (
-        <CDEditModal 
-          cd={editingCD}
-          onSave={handleUpdateCD}
-          onCancel={() => setEditingCD(null)}
-          availableImages={publicImages}
-        />
-      )}
-
-      {/* 编辑产品页面模态框 */}
-      {editingProductPage && (
-        <ProductPageEditModal 
-          page={editingProductPage}
-          onSave={handleUpdateProductPage}
-          onCancel={() => setEditingProductPage(null)}
-          availableImages={publicImages}
-        />
-      )}
-    </div>
-  )
-}
-
-// 商品编辑模态框
-function ProductEditModal({ product, onSave, onCancel, availableImages }) {
-  const [formData, setFormData] = useState({
-    title: product.title || '',
-    description: product.description || '',
-    detailedDescription: product.detailedDescription || '', // 新增详细描述
-    productInfo: product.productInfo || '', // 新增产品信息
-    image: product.image || '',
-    price: product.price || 0,
-    category: product.category || 'music',
-    visible: product.visible || false,
-    order: product.order || 1
-  })
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    onSave(product.id, formData)
-  }
-
-  return (
-    <div style={modalOverlayStyle}>
-      <div style={modalStyle}>
-        <h2>编辑商品</h2>
-        <form onSubmit={handleSubmit}>
-          <div style={formGroupStyle}>
-            <label>商品名称:</label>
-                    <input
-                      type="text"
-              value={formData.title}
-              onChange={(e) => setFormData({...formData, title: e.target.value})}
-              style={modalInputStyle}
-                    />
-                  </div>
-
-          <div style={formGroupStyle}>
-            <label>简短描述:</label>
-                    <textarea
-              value={formData.description}
-              onChange={(e) => setFormData({...formData, description: e.target.value})}
-              style={modalTextareaStyle}
-              rows={2}
-              placeholder="商品的简短描述"
-                    />
-                  </div>
-
-          <div style={formGroupStyle}>
-            <label>详细介绍:</label>
-            <textarea 
-              value={formData.detailedDescription}
-              onChange={(e) => setFormData({...formData, detailedDescription: e.target.value})}
-              style={modalTextareaStyle}
-              rows={5}
-              placeholder="商品的详细介绍，支持多行文本"
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>产品信息:</label>
-            <textarea 
-              value={formData.productInfo}
-              onChange={(e) => setFormData({...formData, productInfo: e.target.value})}
-              style={modalTextareaStyle}
-              rows={3}
-              placeholder="发行信息、规格参数等"
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>商品图片:</label>
-            <select
-              value={formData.image}
-              onChange={(e) => setFormData({...formData, image: e.target.value})}
-              style={modalSelectStyle}
-            >
-              <option value="">选择图片...</option>
-              {availableImages.map(imageUrl => (
-                <option key={imageUrl} value={imageUrl}>
-                  {imageUrl}
-                </option>
-              ))}
-            </select>
-            {formData.image && (
-              <img src={formData.image} alt="预览" style={previewImageStyle} />
-            )}
-          </div>
-
-          <div style={formGroupStyle}>
-                    <label>价格:</label>
-                    <input
-                      type="number"
-              value={formData.price}
-              onChange={(e) => setFormData({...formData, price: parseFloat(e.target.value) || 0})}
-              style={modalInputStyle}
-                    />
-                  </div>
-
-          <div style={formGroupStyle}>
-            <label>
-              <input 
-                type="checkbox"
-                checked={formData.visible}
-                onChange={(e) => setFormData({...formData, visible: e.target.checked})}
-              />
-              在商店中显示
-            </label>
-          </div>
-
-          <div style={modalActionsStyle}>
-            <button type="submit" style={saveButtonStyle}>保存</button>
-            <button type="button" onClick={onCancel} style={cancelButtonStyle}>取消</button>
-          </div>
-        </form>
-      </div>
-    </div>
-  )
-}
-
-// CD编辑模态框
-function CDEditModal({ cd, onSave, onCancel, availableImages }) {
-  const [formData, setFormData] = useState({
-    title: cd.title || '',
-    image: cd.image || '',
-    productLink: cd.productLink || '',
-    visible: cd.visible !== false,
-    order: cd.order || 1
-  })
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    onSave(cd.id, formData)
-  }
-
-  return (
-    <div style={modalOverlayStyle}>
-      <div style={modalStyle}>
-        <h2>编辑CD</h2>
-        <form onSubmit={handleSubmit}>
-          <div style={formGroupStyle}>
-            <label>CD标题:</label>
-            <input 
-              type="text"
-              value={formData.title}
-              onChange={(e) => setFormData({...formData, title: e.target.value})}
-              style={modalInputStyle}
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>CD图片:</label>
-                    <select
-              value={formData.image}
-              onChange={(e) => setFormData({...formData, image: e.target.value})}
-              style={modalSelectStyle}
-                    >
-              <option value="">选择图片...</option>
-              {availableImages.map(imageUrl => (
-                <option key={imageUrl} value={imageUrl}>
-                  {imageUrl}
-                </option>
-              ))}
-                    </select>
-            {formData.image && (
-              <img src={formData.image} alt="预览" style={previewImageStyle} />
-            )}
-                  </div>
-
-          <div style={formGroupStyle}>
-            <label>产品链接:</label>
-            <input 
-              type="text"
-              value={formData.productLink}
-              onChange={(e) => setFormData({...formData, productLink: e.target.value})}
-              style={modalInputStyle}
-              placeholder="/products/album-name"
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-                    <label>
-                      <input
-                        type="checkbox"
-                checked={formData.visible}
-                onChange={(e) => setFormData({...formData, visible: e.target.checked})}
-                      />
-              在轮播中显示
-                    </label>
-                </div>
-
-          <div style={modalActionsStyle}>
-            <button type="submit" style={saveButtonStyle}>保存</button>
-            <button type="button" onClick={onCancel} style={cancelButtonStyle}>取消</button>
-              </div>
-        </form>
-          </div>
-      </div>
-  )
-}
-
-// 样式定义
-const authContainerStyle = {
-  minHeight: '100vh',
-  background: 'linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%)',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  fontFamily: 'monospace'
-}
-
-const authFormStyle = {
-  background: 'rgba(0, 0, 0, 0.8)',
-  padding: '2rem',
-  borderRadius: '8px',
-  minWidth: '300px',
-  textAlign: 'center'
-}
-
-const containerStyle = {
-  minHeight: '100vh',
-  background: 'linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%)',
-  color: 'white',
-  fontFamily: 'monospace',
-  padding: '1rem'
-}
-
-const headerStyle = {
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  marginBottom: '2rem',
-  padding: '1rem',
-  background: 'rgba(0, 0, 0, 0.3)',
-  borderRadius: '8px'
-}
-
-const headerActionsStyle = {
-  display: 'flex',
-  gap: '1rem'
-}
-
-const titleStyle = {
-  color: 'red',
-  fontSize: '1.5rem',
-  margin: 0
-}
-
-const tabsStyle = {
-  display: 'flex',
-  gap: '1rem',
-  marginBottom: '2rem',
-  borderBottom: '1px solid #333'
-}
-
-const tabStyle = {
-  padding: '0.5rem 1rem',
-  background: 'transparent',
-  borderTop: 'none',
-  borderLeft: 'none',
-  borderRight: 'none',
-  borderBottom: 'none',
-  color: '#888',
-  cursor: 'pointer',
-  fontSize: '1rem',
-  fontFamily: 'monospace'
-}
-
-const activeTabStyle = {
-  ...tabStyle,
-  color: 'red',
-  borderBottom: '2px solid red'
-}
-
-const contentStyle = {
-  padding: '1rem'
-}
-
-const sectionHeaderStyle = {
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  marginBottom: '1rem'
-}
-
-const gridStyle = {
-  display: 'grid',
-  gridTemplateColumns: 'repeat(auto-fill, minmax(250px, 1fr))',
-  gap: '1rem'
-}
-
-const imageGridStyle = {
-  display: 'grid',
-  gridTemplateColumns: 'repeat(auto-fill, minmax(200px, 1fr))',
-  gap: '1rem',
-  marginBottom: '2rem'
-}
-
-const cardStyle = {
-  background: 'rgba(0, 0, 0, 0.5)',
-  borderRadius: '8px',
-  overflow: 'hidden',
-  border: '1px solid #333'
-}
-
-const imageCardStyle = {
-  background: 'rgba(0, 0, 0, 0.5)',
-  borderRadius: '8px',
-  padding: '1rem',
-  textAlign: 'center',
-  border: '1px solid #333'
-}
-
-const cardImageStyle = {
-  width: '100%',
-  height: '150px',
-  objectFit: 'cover'
-}
-
-const imageStyle = {
-  width: '100%',
-  height: '120px',
-  objectFit: 'cover',
-  borderRadius: '4px'
-}
-
-const cardContentStyle = {
-  padding: '1rem'
-}
-
-const cardActionsStyle = {
-  display: 'flex',
-  gap: '0.5rem',
-  marginTop: '1rem'
-}
-
-const buttonStyle = {
-  padding: '0.5rem 1rem',
-  background: 'red',
-  color: 'white',
-  borderTop: 'none',
-  borderLeft: 'none',
-  borderRight: 'none',
-  borderBottom: 'none',
-  borderRadius: '4px',
-  cursor: 'pointer',
-  fontFamily: 'monospace'
-}
-
-const createButtonStyle = {
-  ...buttonStyle,
-  background: '#28a745'
-}
-
-const editButtonStyle = {
-  ...buttonStyle,
-  background: '#ffc107',
-  color: 'black'
-}
-
-const deleteButtonStyle = {
-  ...buttonStyle,
-  background: '#dc3545'
-}
-
-const uploadButtonStyle = {
-  ...buttonStyle,
-  background: '#17a2b8',
-  cursor: 'pointer'
-}
-
-const copyButtonStyle = {
-  ...buttonStyle,
-  background: '#6c757d',
-  fontSize: '0.8rem',
-  marginTop: '0.5rem'
-}
-
-const backButtonStyle = {
-  ...buttonStyle,
-  background: '#6c757d'
-}
-
-const logoutButtonStyle = {
-  ...buttonStyle,
-  background: '#dc3545'
-}
-
-const formStyle = {
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '1rem'
-}
-
-const inputStyle = {
-  padding: '0.5rem',
-  borderRadius: '4px',
-  border: '1px solid #333',
-  background: 'rgba(0, 0, 0, 0.7)',
-  color: 'white',
-  fontFamily: 'monospace'
-}
-
-const errorStyle = {
-  color: 'red',
-  fontSize: '0.9rem',
-  marginTop: '0.5rem'
-}
-
-const errorBannerStyle = {
-  background: '#dc3545',
-  color: 'white',
-  padding: '0.5rem',
-  borderRadius: '4px',
-  marginBottom: '1rem'
-}
-
-const loadingBannerStyle = {
-  background: '#ffc107',
-  color: 'black',
-  padding: '0.5rem',
-  borderRadius: '4px',
-  marginBottom: '1rem',
-  textAlign: 'center'
-}
-
-// 模态框样式
-const modalOverlayStyle = {
-  position: 'fixed',
-  top: 0,
-  left: 0,
-  right: 0,
-  bottom: 0,
-  background: 'rgba(0, 0, 0, 0.8)',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  zIndex: 1000
-}
-
-const modalStyle = {
-  background: '#2d2d2d',
-  padding: '2rem',
-  borderRadius: '8px',
-  width: '90%',
-  maxWidth: '500px',
-  maxHeight: '80vh',
-  overflow: 'auto'
-}
-
-const formGroupStyle = {
-  marginBottom: '1rem'
-}
-
-const modalInputStyle = {
-  ...inputStyle,
-  width: '100%',
-  marginTop: '0.5rem'
-}
-
-const modalTextareaStyle = {
-  ...modalInputStyle,
-  resize: 'vertical'
-}
-
-const modalSelectStyle = {
-  ...modalInputStyle
-}
-
-const previewImageStyle = {
-  width: '100%',
-  maxHeight: '150px',
-  objectFit: 'cover',
-  marginTop: '0.5rem',
-  borderRadius: '4px'
-}
-
-const modalActionsStyle = {
-  display: 'flex',
-  gap: '1rem',
-  justifyContent: 'flex-end',
-  marginTop: '1rem'
-}
-
-const saveButtonStyle = {
-  ...buttonStyle,
-  background: '#28a745'
-}
-
-const cancelButtonStyle = {
-  ...buttonStyle,
-  background: '#6c757d'
-}
-
-// 产品页面编辑模态框组件
-function ProductPageEditModal({ page, onSave, onCancel, availableImages }) {
-  const [formData, setFormData] = useState({
-    title: page.title || '',
-    description: page.description || '',
-    image: page.image || '',
-    price: page.price || '',
-    trackList: page.trackList ? page.trackList.join('\n') : '',
-    catalog: page.details?.catalog || '',
-    album: page.details?.album || '',
-    releaseType: page.details?.releaseType || '',
-    releaseDate: page.details?.releaseDate || '',
-    label: page.details?.label || '',
-    ar: page.details?.ar || '',
-    writer: page.details?.writer || '',
-    producer: page.details?.producer || '',
-    mixing: page.details?.mixing || '',
-    master: page.details?.master || '',
-    creativeDirector: page.details?.creativeDirector || '',
-    artDirector: page.details?.artDirector || ''
-  })
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    
-    const pageData = {
-      title: formData.title,
-      description: formData.description,
-      image: formData.image,
-      price: formData.price,
-      trackList: formData.trackList.split('\n').filter(track => track.trim()),
-      details: {
-        catalog: formData.catalog,
-        album: formData.album,
-        releaseType: formData.releaseType,
-        releaseDate: formData.releaseDate,
-        label: formData.label,
-        ar: formData.ar,
-        writer: formData.writer,
-        producer: formData.producer,
-        mixing: formData.mixing,
-        master: formData.master,
-        creativeDirector: formData.creativeDirector,
-        artDirector: formData.artDirector
-      }
-    }
-    
-    onSave(page.id, pageData)
-  }
-
-  return (
-    <div style={modalOverlayStyle} onClick={onCancel}>
-      <div style={{ ...modalStyle, maxWidth: '700px' }} onClick={e => e.stopPropagation()}>
-        <h2>编辑产品页面: {page.title}</h2>
-        <form onSubmit={handleSubmit}>
-          <div style={formGroupStyle}>
-            <label>产品标题</label>
-            <input
-              type="text"
-              value={formData.title}
-              onChange={e => setFormData({...formData, title: e.target.value})}
-              style={modalInputStyle}
-              required
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>产品描述</label>
-            <textarea
-              value={formData.description}
-              onChange={e => setFormData({...formData, description: e.target.value})}
-              style={modalTextareaStyle}
-              rows="3"
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>产品图片</label>
-            <select
-              value={formData.image}
-              onChange={e => setFormData({...formData, image: e.target.value})}
-              style={modalSelectStyle}
-            >
-              <option value="">选择图片</option>
-              {availableImages.map(image => (
-                <option key={image} value={image}>{image}</option>
-              ))}
-            </select>
-            {formData.image && (
-              <img src={formData.image} alt="预览" style={previewImageStyle} />
-            )}
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>价格</label>
-            <input
-              type="text"
-              value={formData.price}
-              onChange={e => setFormData({...formData, price: e.target.value})}
-              style={modalInputStyle}
-              placeholder="€23.00"
-            />
-          </div>
-
-          <div style={formGroupStyle}>
-            <label>曲目列表 (每行一个曲目)</label>
-            <textarea
-              value={formData.trackList}
-              onChange={e => setFormData({...formData, trackList: e.target.value})}
-              style={modalTextareaStyle}
-              rows="8"
-              placeholder="A1 · TRACK ONE&#10;A2 · TRACK TWO&#10;B1 · TRACK THREE"
-            />
-          </div>
-
-          <h3>发布详情</h3>
-          
-          <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: '1rem' }}>
-            <div style={formGroupStyle}>
-              <label>目录号</label>
-              <input
-                type="text"
-                value={formData.catalog}
-                onChange={e => setFormData({...formData, catalog: e.target.value})}
-                style={modalInputStyle}
-                placeholder="YR0189"
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>专辑名称</label>
-              <input
-                type="text"
-                value={formData.album}
-                onChange={e => setFormData({...formData, album: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>发布类型</label>
-              <input
-                type="text"
-                value={formData.releaseType}
-                onChange={e => setFormData({...formData, releaseType: e.target.value})}
-                style={modalInputStyle}
-                placeholder="Album / EP"
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>发布日期</label>
-              <input
-                type="text"
-                value={formData.releaseDate}
-                onChange={e => setFormData({...formData, releaseDate: e.target.value})}
-                style={modalInputStyle}
-                placeholder="20/09/2024"
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>厂牌</label>
-              <input
-                type="text"
-                value={formData.label}
-                onChange={e => setFormData({...formData, label: e.target.value})}
-                style={modalInputStyle}
-                placeholder="YEAR0001"
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>A&R</label>
-              <input
-                type="text"
-                value={formData.ar}
-                onChange={e => setFormData({...formData, ar: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>作词人</label>
-              <input
-                type="text"
-                value={formData.writer}
-                onChange={e => setFormData({...formData, writer: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>制作人</label>
-              <input
-                type="text"
-                value={formData.producer}
-                onChange={e => setFormData({...formData, producer: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>混音师</label>
-              <input
-                type="text"
-                value={formData.mixing}
-                onChange={e => setFormData({...formData, mixing: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>母带工程师</label>
-              <input
-                type="text"
-                value={formData.master}
-                onChange={e => setFormData({...formData, master: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>创意总监</label>
-              <input
-                type="text"
-                value={formData.creativeDirector}
-                onChange={e => setFormData({...formData, creativeDirector: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-
-            <div style={formGroupStyle}>
-              <label>艺术总监</label>
-              <input
-                type="text"
-                value={formData.artDirector}
-                onChange={e => setFormData({...formData, artDirector: e.target.value})}
-                style={modalInputStyle}
-              />
-            </div>
-          </div>
-
-          <div style={modalActionsStyle}>
-            <button type="button" onClick={onCancel} style={cancelButtonStyle}>
-              取消
-            </button>
-            <button type="submit" style={saveButtonStyle}>
-              保存更改
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  )
-}
